requirements.txt
Cython==0.20.2
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Login==0.2.11
Flask-Migrate==1.2.0
Flask-Permissions==0.2.3
Flask-Principal==0.4.0
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Flask-WTF==0.10.1
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.7
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
alembic==0.6.6
blinker==1.3
certifi==0.0.8
chardet==1.0.1
decorator==3.4.0
gunicorn==0.17.2
itsdangerous==0.24
pbr==0.10.0
psycopg2==2.5.3
requests==0.11.1
six==1.7.3
sqlalchemy-migrate==0.9.1
validators==0.6.0
wsgiref==0.1.2
